backend/requirements.txt
# Core FastAPI and server dependencies
fastapi==0.110.1
uvicorn==0.25.0
starlette==0.37.2
pydantic==2.12.4
pydantic_core==2.41.5

# Database
motor==3.3.1
pymongo==4.5.0
dnspython==2.8.0

# Authentication & Security
bcrypt==4.1.3
python-jose==3.5.0
PyJWT==2.10.1
passlib==1.7.4
python-multipart==0.0.20

# Google OAuth
google-auth==2.43.0
google-auth-oauthlib==1.2.1

# HTTP Client
httpx==0.28.1
httpcore==1.0.9
requests==2.32.5
aiohttp==3.13.2

# OpenAI
openai==1.99.9
tiktoken==0.12.0

# Stripe
stripe==14.0.1

# Document Generation
python-docx==1.1.2
reportlab==4.4.5

# Utilities
python-dotenv==1.2.1
email-validator==2.3.0
python-dateutil==2.9.0.post0
pytz==2025.2

# Data processing
pandas==2.3.3
numpy==2.3.5

# Async support
anyio==4.11.0
sniffio==1.3.1

# Core dependencies
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
urllib3==2.5.0
six==1.17.0
packaging==25.0

# Additional utilities
tenacity==9.1.2
rich==14.2.0
PyYAML==6.0.3

# Type hints
typing_extensions==4.15.0

# MongoDB connection
aiohappyeyeballs==2.6.1
